Add functional Bank
Functional application for counting how may money is on different
accounts using interfaces.

+++ 3.2_Interfaces/src/model/interfaces/Account.java
@@ -6,6 +6,6 @@ package model.interfaces;
 public interface Account {
     void gain();
     void interest();
-    int getAmount();
+    double getAmount();
     String getName();
 }
